fix(#267): cover the auxiliary MCP bridge; move mcpContentToText to src/mcp/
Review caught a third bridge my fix missed: composeAuxiliaryBridge
(src/mcp/compose-auxiliary.ts) still built McpCallResult.text with
`typeof content === 'string' ? content : JSON.stringify(content)`, so an
auxiliary tool returning the standard MCP text-block array still handed the LLM
the raw `[{"type":"text","text":"…"}]` envelope. Wired mcpContentToText() there
too.

src/mcp/ does not (and should not) depend on src/smart-agent/, so the shared
helper moved from src/smart-agent/mcp/ to src/mcp/ (its natural home next to
compose-auxiliary); all three bridges import it from there — buildMcpBridge,
buildNamespacedMcpBridge, composeAuxiliaryBridge.

Test: compose-auxiliary.test.ts gains a bridge-level case (text-block content
→ unwrapped, not raw JSON); the existing structured-object case (→ JSON.stringify)
still passes, so the conservative boundary is pinned at the bridge too.

mcp dir suites 21/21; smart-agent mcp + controller + bridge suites 362/362;
`biome check packages` (lint:check) exit 0; tsc clean.
