+++ packages/llm-agent-server-libs/src/mcp/mcp-content.ts
