+++ packages/llm-agent-server-libs/src/smart-agent/mcp/namespaced-bridge.ts
@@ -6,7 +6,7 @@ import type {
 } from '@mcp-abap-adt/llm-agent';
 import { bindToolCallName } from '@mcp-abap-adt/llm-agent';
 import { DefaultMcpFailureClassifier } from '@mcp-abap-adt/llm-agent-mcp';
-import { mcpContentToText } from './mcp-content.js';
+import { mcpContentToText } from '../../mcp/mcp-content.js';
 
 /**
  * Resolve the client that currently owns a given provenance `slotIndex`: the
